Moduleize, Break into files, Refactor for Client+Server compatibility
