+++ lib/ts3query/errors.rb
@@ -1,0 +1,3 @@
+module TS3Query
+  class ConnectionRefused < StandardError; end
+end
